Sort out enemy cleanup when they die.

+++ RougeLike/Source/Characters/Enemies/EnemyStates/EnemyDead.h
@@ -19,5 +19,5 @@ public:
 	const Type type() const { return Type::Dead; }
 
 private:
-	Uint8 mAlpha;
+	double mAlpha;
 };
